pyproject.toml
[tool.poetry]
name = "automate-content-search-cnu"
version = "0.1.1"
description = "Automate content search for CNU"
authors = ["davirolim <davirolim94@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "^0.32.2"}
# crewai = "^0.16.0"
pydantic = "^2.6.3"
load-dotenv = "^0.1.0"
crewai-tools = "^0.3.0"
fastapi = "^0.111.0"
uvicorn = "0.30.1"
asyncpg = "0.29.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"